# plan_vision_backend/api-docs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/styles": {
      "post": {
        "operationId": "StyleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStyleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Style created."
          }
        },
        "summary": "Create a new style entry.",
        "tags": [
          "Styles"
        ]
      },
      "get": {
        "operationId": "StyleController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of styles."
          }
        },
        "summary": "List all styles alphabetically.",
        "tags": [
          "Styles"
        ]
      }
    },
    "/styles/{id}": {
      "get": {
        "operationId": "StyleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Style details."
          },
          "404": {
            "description": "Style not found."
          }
        },
        "summary": "Get a single style by ID.",
        "tags": [
          "Styles"
        ]
      },
      "patch": {
        "operationId": "StyleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStyleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Style updated."
          },
          "404": {
            "description": "Style not found."
          }
        },
        "summary": "Update style fields.",
        "tags": [
          "Styles"
        ]
      },
      "delete": {
        "operationId": "StyleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Style deleted."
          },
          "404": {
            "description": "Style not found."
          }
        },
        "summary": "Delete a style.",
        "tags": [
          "Styles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Plan Vision API",
    "description": "REST API for Plan Vision backend (users, auth, providers).",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_session",
        "description": "Better Auth session cookie"
      },
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Optional Better Auth access token"
      }
    },
    "schemas": {
...
      "CreateStyleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User-facing name (e.g., \"Modern\").",
            "maxLength": 255
          },
          "thumbnailUrl": {
            "type": "string",
            "description": "Thumbnail URL displayed in the UI."
          },
          "promptFragment": {
            "type": "string",
            "description": "Prompt fragment sent to the AI when this style is chosen."
          }
        },
        "required": [
          "name",
          "thumbnailUrl",
          "promptFragment"
        ]
      },
      "UpdateStyleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User-facing name (e.g., \"Modern\").",
            "maxLength": 255
          },
          "thumbnailUrl": {
            "type": "string",
            "description": "Thumbnail URL displayed in the UI."
          },
          "promptFragment": {
            "type": "string",
            "description": "Prompt fragment sent to the AI when this style is chosen."
          }
        }
      },
...
    }
  }
}